recognition/StyleGAN_47462153/requirements.txt
torch
torchvision
Pillow
numpy
matplotlib
scikit-learn
